pyproject.toml
[project]
name = "restgdfweb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.111.0",
    "jinja2==3.1.4",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "bmi-arcgis-restapi>=2.4.9",
    "geopandas>=1.0.1",
    "openlayer>=0.2.0a42",
]
